backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis and Caching
redis<5.0.0,>=4.5.2
hiredis==2.2.3

# Background Tasks
celery[redis]==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Core ML Libraries (Python 3.12 compatible)
scikit-learn>=1.3.2
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Collaborative Filtering (Python 3.12 compatible)
scikit-surprise>=1.1.3

# Deep Learning & NLP (Latest versions for Python 3.12)
torch>=2.2.0
transformers>=4.35.0
sentence-transformers>=2.2.0
textblob==0.17.1

# Additional ML utilities
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru==0.7.2

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0